requirements.txt
streamlit
langchain_google_genai
PyPDF2
langchain-community
PIL
